fix a lighting bug
Placing a light source close to a weaker light source and then breaking the stronger source damaged the weaker source's output. The specific trigger is the stronger source setting a light component at the
weaker source to be higher than the weaker source's output. I based my light subtraction algorithm on Seed of Andromeda's. I guess that they never encountered this issue because their light sources are
always the same intensity with the only difference being the color.